chore: move LeftSidebar and RightSidebar to /layout folder
